+++ pkg/app/mutate_app.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/Masterminds/semver/v3"
 	"github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	"github.com/giantswarm/app/v6/pkg/key"
 	"github.com/giantswarm/k8sclient/v6/pkg/k8sclient"
@@ -125,34 +124,12 @@ func (m *Mutator) MutateApp(ctx context.Context, oldApp, app v1alpha1.App, opera
 		}
 	}
 
-	var patchLabels bool
-
 	labelPatches, err := m.mutateLabels(ctx, app, appVersionLabel)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 	if len(labelPatches) > 0 {
 		result = append(result, labelPatches...)
-		patchLabels = true
-	}
-
-	ver, err := semver.NewVersion(appVersionLabel)
-	if !isManagedInOrg && err != nil {
-		m.logger.Debugf(ctx, "skipping mutation of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, appVersionLabel)
-		return nil, nil
-	}
-
-	// If the app CR does not have the unique version and is < 3.0.0 we skip
-	// the defaulting logic apart from the labels. This is so the admission
-	// controller is not enabled for existing platform releases.
-	if !isManagedInOrg && key.VersionLabel(app) != uniqueAppCRVersion && ver.Major() < 3 {
-		if patchLabels {
-			m.logger.Debugf(ctx, "mutating only labels of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, appVersionLabel)
-			return result, nil
-		}
-
-		m.logger.Debugf(ctx, "skipping mutation of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, appVersionLabel)
-		return nil, nil
 	}
 
 	configPatches, err := m.mutateConfig(ctx, app)