+++ pkg/app/validate_app.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/Masterminds/semver/v3"
 	"github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	"github.com/giantswarm/app/v6/pkg/key"
 	"github.com/giantswarm/app/v6/pkg/validation"
@@ -111,22 +110,6 @@ func (v *Validator) Validate(request *admissionv1.AdmissionRequest) (bool, error
 
 	if request.Operation == admissionv1.Update && !app.DeletionTimestamp.IsZero() {
 		v.logger.Debugf(ctx, "skipping validation for UPDATE operation of app %#q in namespace %#q with non-zero deletion timestamp", app.Name, app.Namespace)
-		return true, nil
-	}
-
-	isManagedInOrg := !key.InCluster(app) && key.IsInOrgNamespace(app)
-
-	ver, err := semver.NewVersion(key.VersionLabel(app))
-	if !isManagedInOrg && err != nil {
-		v.logger.Debugf(ctx, "skipping validation of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, key.VersionLabel(app))
-		return true, nil
-	}
-
-	// If the app CR does not have the unique version and is < 3.0.0 we skip
-	// the validation logic. This is so the admission controller is not
-	// enabled for existing platform releases.
-	if !isManagedInOrg && key.VersionLabel(app) != uniqueAppCRVersion && ver.Major() < 3 {
-		v.logger.Debugf(ctx, "skipping validation of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, key.VersionLabel(app))
 		return true, nil
 	}