+++ pkg/app/validate_app_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/giantswarm/k8sclient/v6/pkg/k8sclienttest"
 	"github.com/giantswarm/micrologger/microloggertest"
 	admissionv1 "k8s.io/api/admission/v1"
+	authv1 "k8s.io/api/authentication/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -16,9 +17,40 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake" //nolint:staticcheck
 
 	"github.com/giantswarm/app-admission-controller/internal/recorder"
+	secins "github.com/giantswarm/app-admission-controller/internal/security/inspector"
 )
 
 func Test_ValidateApp(t *testing.T) {
+	secInsCfg := secins.Config{
+		Logger: microloggertest.New(),
+
+		NamespaceBlacklist: []string{
+			"capi-",
+			"-prometheus",
+			"draughtsman",
+			"flux-giantswarm",
+			"giantswarm",
+			"kube-system",
+			"monitoring",
+		},
+		GroupWhitelist: []string{
+			"giantswarm:giantswarm:giantswarm-admins",
+		},
+		UserWhitelist: []string{
+			"system:serviceaccount:draughtsman:",
+			"system:serviceaccount:giantswarm:",
+			"system:serviceaccount:flux-giantswarm:",
+			"system:serviceaccount:kube-system:",
+		},
+		AppBlacklist: []string{
+			"app-operator",
+		},
+		CatalogBlacklist: []string{
+			"control-plane-catalog",
+			"control-plane-test-catalog",
+		},
+	}
+
 	tests := []struct {
 		name        string
 		obj         *admissionv1.AdmissionRequest
@@ -211,6 +243,431 @@ func Test_ValidateApp(t *testing.T) {
 			},
 			expectedErr: "validation error: label `giantswarm.io/cluster` not found",
 		},
+		{
+			name: "flawless unique app validation",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "hello-world",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "giantswarm",
+    							"name": "hello-world",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "demo0-cluster-values",
+										"namespace": "demo0"
+									}
+								},
+    							"kubeConfig": {
+									"inCluster": true
+								},
+								"userConfig": {
+									"configMap": {
+										"name": "hello-world-user-values",
+										"namespace": "demo0"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "00001@customer.onmicrosoft.com",
+					Groups: []string{
+						"customer:CUSTOM-GROUP",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("giantswarm", "default"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("demo0-cluster-values", "demo0"),
+				newTestConfigMap("hello-world-user-values", "demo0"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("demo0-kubeconfig", "demo0"),
+			},
+		},
+		{
+			name: "referencing protected configuration as regular user",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "hello-world",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "giantswarm",
+    							"name": "hello-world",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "demo0-cluster-values",
+										"namespace": "demo0"
+									}
+								},
+    							"kubeConfig": {
+									"context": {
+										"name": "demo0-kubeconfig"
+									},
+									"inCluster": false,
+									"secret": {
+										"name": "demo0-kubeconfig",
+										"namespace": "demo0"
+									}
+								},
+								"userConfig": {
+									"secret": {
+										"name": "vault-token",
+										"namespace": "giantswarm"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "00001@customer.onmicrosoft.com",
+					Groups: []string{
+						"customer:CUSTOMER-GROUP",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("giantswarm", "default"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("demo0-cluster-values", "demo0"),
+				newTestConfigMap("hello-world-user-values", "demo0"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("demo0-kubeconfig", "demo0"),
+				newTestSecret("vault-token", "giantswarm"),
+			},
+			expectedErr: "security violation error: references to `giantswarm` namespace not allowed",
+		},
+		{
+			name: "referencing protected configuration as regular service account",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "hello-world",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "giantswarm",
+    							"name": "hello-world",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "demo0-cluster-values",
+										"namespace": "demo0"
+									}
+								},
+    							"kubeConfig": {
+									"context": {
+										"name": "demo0-kubeconfig"
+									},
+									"inCluster": false,
+									"secret": {
+										"name": "demo0-kubeconfig",
+										"namespace": "demo0"
+									}
+								},
+								"userConfig": {
+									"configMap": {
+										"name": "capi-credentials",
+										"namespace": "capi-system"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "system:serviceaccount:default:automation",
+					Groups: []string{
+						"system:authenticated",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("giantswarm", "default"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("demo0-cluster-values", "demo0"),
+				newTestConfigMap("hello-world-user-values", "demo0"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("demo0-kubeconfig", "demo0"),
+				newTestSecret("vault-token", "giantswarm"),
+			},
+			expectedErr: "security violation error: references to `capi-system` namespace not allowed",
+		},
+		{
+			name: "referencing protected configuration as privileged service account",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "app-operator-demo0",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "control-plane",
+    							"name": "app-operator",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "app-operator-cluster-values",
+										"namespace": "giantswarm"
+									}
+								},
+    							"kubeConfig": {
+									"inCluster": true
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "system:serviceaccount:giantswarm:cluster-operator-3-13-0",
+					Groups: []string{
+						"system:authenticated",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("control-plane", "giantswarm"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("app-operator-cluster-values", "giantswarm"),
+			},
+		},
+		{
+			name: "install blacklisted app as regular service account",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "app-operator-new",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "control-plane-catalog",
+    							"name": "app-operator",
+    							"namespace": "demo0",
+    							"kubeConfig": {
+									"inCluster": true
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "system:serviceaccount:default:automation",
+					Groups: []string{
+						"system:authenticated",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("control-plane", "giantswarm"),
+			},
+			expectedErr: "security violation error: installing `app-operator` from `control-plane-catalog` catalog is not allowed",
+		},
+		{
+			name: "modify existing App CR as a regular user",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "UPDATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "app-operator-demo0",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "control-plane-catalog",
+    							"name": "app-operator",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "app-operator-cluster-values",
+										"namespace": "giantswarm"
+									}
+								},
+    							"kubeConfig": {
+									"inCluster": true
+								},
+								"userConfig": {
+									"secret": {
+										"name": "vault-token",
+										"namespace": "giantswarm"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "00001@customer.onmicrosoft.com",
+					Groups: []string{
+						"customer:CUSTOMER-GROUP",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("control-plane", "giantswarm"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("app-operator-cluster-values", "giantswarm"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("vault-token", "giantswarm"),
+			},
+			expectedErr: "security violation error: installing `app-operator` from `control-plane-catalog` catalog is not allowed",
+		},
+		{
+			name: "modify existing App CR as a GS member",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "UPDATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "app-operator-demo0",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "control-plane",
+    							"name": "app-operator",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "app-operator-cluster-values",
+										"namespace": "giantswarm"
+									}
+								},
+    							"kubeConfig": {
+									"inCluster": true
+								},
+								"userConfig": {
+									"secret": {
+										"name": "app-operator-secrets",
+										"namespace": "giantswarm"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				OldObject: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "app-operator-demo0",
+    							"namespace": "demo0",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "control-plane",
+    							"name": "app-operator",
+    							"namespace": "demo0",
+    							"config": {
+									"configMap": {
+										"name": "app-operator-cluster-values",
+										"namespace": "giantswarm"
+									}
+								},
+    							"kubeConfig": {
+									"inCluster": true
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "user@giantswarm.io",
+					Groups: []string{
+						"giantswarm:giantswarm:giantswarm-admins",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("control-plane", "giantswarm"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("app-operator-cluster-values", "giantswarm"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("app-operator-secrets", "giantswarm"),
+			},
+		},
 	}
 
 	for i, tc := range tests {
@@ -252,11 +709,17 @@ func Test_ValidateApp(t *testing.T) {
 				event = recorder.New(c)
 			}
 
+			ins, err := secins.New(secInsCfg)
+			if err != nil {
+				t.Fatalf("error == %#v, want nil", err)
+			}
+
 			c := ValidatorConfig{
 				Event:     event,
 				K8sClient: k8sClient,
 				Logger:    microloggertest.New(),
 				Provider:  "aws",
+				Inspector: ins,
 			}
 
 			r, err := NewValidator(c)