+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Run additional validation logic for submitting the `0.0.0`-labeld App CRs by a non privileged users.
+
 ## [0.16.3] - 2022-02-25
 
 ### Fixed